# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Соберёмся API",
    "version": "0.4.0",
    "description": "Бот и mini app используют общие встречи. Геопозиция и поиск поддерживают города России, реальные места рядом загружаются из OpenStreetMap."
  },
  "servers": [
    {
      "url": "https://soberemsya.139.100.225.250.sslip.io",
      "description": "Public hackathon MVP"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local demo"
    }
  ],
  "paths": {
...
    "/api/config": {
      "get": {
        "summary": "Public configuration",
        "operationId": "get__api_config",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "allowDemo",
                    "botUsername",
                    "appUrl",
                    "cities"
                  ],
                  "properties": {
                    "allowDemo": {
                      "type": "boolean"
                    },
                    "botUsername": {
                      "type": "string"
                    },
                    "appUrl": {
                      "type": "string"
                    },
                    "cities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/City"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog": {
      "get": {
        "summary": "Prepared catalogue",
        "operationId": "get__api_catalog",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "activities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Activity"
                      }
                    }
                  },
                  "required": [
                    "activities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/demo": {
      "post": {
        "summary": "Create isolated demo identity",
        "operationId": "post__api_auth_demo",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/Identity"
                    }
                  },
                  "required": [
                    "token",
                    "user"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  }
                },
                "required": [
                  "displayName"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "summary": "List joined sessions",
        "operationId": "get__api_sessions",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionView"
                      }
                    }
                  },
                  "required": [
                    "sessions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create session",
        "operationId": "post__api_sessions",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "settings": {
                    "$ref": "#/components/schemas/SessionSettings"
                  },
                  "preference": {
                    "$ref": "#/components/schemas/Preference"
                  }
                },
                "required": [
                  "settings",
                  "preference"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 80
            }
          }
        ]
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "summary": "Room and own preference only",
        "operationId": "get__api_sessions_id",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ]
      }
    },
    "/api/sessions/{id}/join": {
      "post": {
        "summary": "Join using invitation identifier",
        "operationId": "post__api_sessions_id_join",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "preference": {
                    "$ref": "#/components/schemas/Preference"
                  }
                },
                "required": [
                  "preference"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/preferences": {
      "put": {
        "summary": "Replace own preference; invalidate previous plan",
        "operationId": "put__api_sessions_id_preferences",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "preference": {
                    "$ref": "#/components/schemas/Preference"
                  }
                },
                "required": [
                  "preference"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/demo-fill": {
      "post": {
        "summary": "Owner of demo room adds four marked demo members",
        "operationId": "post__api_sessions_id_demo-fill",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/plan": {
      "post": {
        "summary": "Calculate group plan; requires two answers",
        "operationId": "post__api_sessions_id_plan",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/share": {
      "post": {
        "summary": "Prepare sharing text and URL; sends no message",
        "operationId": "post__api_sessions_id_share",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "link": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "text",
                    "link"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "400 invalid parameters; 401 invalid auth; 403 forbidden; 404 unknown room; 409 conflict; 410 expired; 422 no match; 429 rate limit; 500 retryable server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{20,40}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/auth/link": {
      "post": {
        "summary": "Exchange private one-use bot login ticket",
        "operationId": "exchangeBotWebTicket",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ticket"
                ],
                "properties": {
                  "ticket": {
                    "type": "string",
                    "minLength": 43,
                    "maxLength": 43
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/Identity"
                    }
                  },
                  "required": [
                    "token",
                    "user"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Ticket expired or already consumed"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Current validated identity",
        "operationId": "currentIdentity",
        "security": [
          {
            "MaxInitData": []
          },
          {
            "DemoBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/Identity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Session invalid or expired"
          }
        }
      }
    },
    "/api/location/resolve": {
      "post": {
        "summary": "Определить город России и точки встречи по геопозиции",
        "operationId": "resolveLocation",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coordinates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Город и точки встречи",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/City"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные координаты"
          },
          "503": {
            "description": "Сервис геоданных временно недоступен"
          }
        }
      }
    },
    "/api/location/search": {
      "get": {
        "summary": "Найти город России по названию",
        "operationId": "searchRussianCities",
        "security": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Найденные города",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cities"
                  ],
                  "properties": {
                    "cities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/City"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Сервис геоданных временно недоступен"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "City": {
        "type": "object",
        "required": [
          "id",
          "name",
          "center",
          "coverageKm",
          "points"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "center": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "coverageKm": {
            "type": "number"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeetingPoint"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "MaxInitData": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Init-Data",
        "description": "Signed original WebApp.initData; 1h max age"
      },
      "DemoBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session from POST /api/auth/demo (demo only) or POST /api/auth/link (verified MAX account); never a bot token."
      }
    }
  }
}